Move array fixer classes to new namespace
