+++ src/Fixer/ArrayNotation/NoWhitespaceBeforeCommaInArrayFixer.php
@@ -10,16 +10,15 @@
  * with this source code in the file LICENSE.
  */
 
-namespace PhpCsFixer\Fixer\Symfony;
+namespace PhpCsFixer\Fixer\ArrayNotation;
 
 use PhpCsFixer\AbstractFixer;
-use PhpCsFixer\Tokenizer\Token;
 use PhpCsFixer\Tokenizer\Tokens;
 
 /**
  * @author Adam Marczuk <adam@marczuk.info>
  */
-final class WhitespaceAfterCommaInArrayFixer extends AbstractFixer
+final class NoWhitespaceBeforeCommaInArrayFixer extends AbstractFixer
 {
     /**
      * {@inheritdoc}
@@ -46,7 +45,7 @@ final class WhitespaceAfterCommaInArrayFixer extends AbstractFixer
      */
     public function getDescription()
     {
-        return 'In array declaration, there MUST be a whitespace after each comma.';
+        return 'In array declaration, there MUST NOT be a whitespace before each comma.';
     }
 
     /**
@@ -67,8 +66,10 @@ final class WhitespaceAfterCommaInArrayFixer extends AbstractFixer
 
         for ($i = $endIndex - 1; $i > $startIndex; --$i) {
             $i = $this->skipNonArrayElements($i, $tokens);
-            if ($tokens[$i]->equals(',') && !$tokens[$i + 1]->isWhitespace()) {
-                $tokens->insertAt($i + 1, new Token(array(T_WHITESPACE, ' ')));
+            $currentToken = $tokens[$i];
+            $prevIndex = $tokens->getPrevNonWhitespace($i - 1);
+            if ($currentToken->equals(',') && !$tokens[$prevIndex]->equals(array(T_END_HEREDOC))) {
+                $tokens->removeLeadingWhitespace($i);
             }
         }
     }
