+++ src/Fixer/ArrayNotation/TrailingCommaInMultilineArrayFixer.php
@@ -10,17 +10,18 @@
  * with this source code in the file LICENSE.
  */
 
-namespace PhpCsFixer\Fixer\Symfony;
+namespace PhpCsFixer\Fixer\ArrayNotation;
 
 use PhpCsFixer\AbstractFixer;
+use PhpCsFixer\Tokenizer\Token;
 use PhpCsFixer\Tokenizer\Tokens;
 use PhpCsFixer\Tokenizer\TokensAnalyzer;
 
 /**
- * @author Dariusz Rumiński <dariusz.ruminski@gmail.com>
  * @author Sebastiaan Stok <s.stok@rollerscapes.net>
+ * @author Dariusz Rumiński <dariusz.ruminski@gmail.com>
  */
-final class NoTrailingCommaInSinglelineArrayFixer extends AbstractFixer
+final class TrailingCommaInMultilineArrayFixer extends AbstractFixer
 {
     /**
      * {@inheritdoc}
@@ -37,7 +38,7 @@ final class NoTrailingCommaInSinglelineArrayFixer extends AbstractFixer
     {
         $tokensAnalyzer = new TokensAnalyzer($tokens);
 
-        for ($index = 0, $c = $tokens->count(); $index < $c; ++$index) {
+        for ($index = $tokens->count() - 1; $index >= 0; --$index) {
             if ($tokensAnalyzer->isArray($index)) {
                 $this->fixArray($tokens, $index);
             }
@@ -49,14 +50,14 @@ final class NoTrailingCommaInSinglelineArrayFixer extends AbstractFixer
      */
     public function getDescription()
     {
-        return 'PHP single-line arrays should not have trailing comma.';
+        return 'PHP multi-line arrays should have a trailing comma.';
     }
 
     private function fixArray(Tokens $tokens, $index)
     {
         $tokensAnalyzer = new TokensAnalyzer($tokens);
 
-        if ($tokensAnalyzer->isArrayMultiLine($index)) {
+        if (!$tokensAnalyzer->isArrayMultiLine($index)) {
             return;
         }
 
@@ -72,9 +73,15 @@ final class NoTrailingCommaInSinglelineArrayFixer extends AbstractFixer
         $beforeEndIndex = $tokens->getPrevMeaningfulToken($endIndex);
         $beforeEndToken = $tokens[$beforeEndIndex];
 
-        if ($beforeEndToken->equals(',')) {
-            $tokens->removeTrailingWhitespace($beforeEndIndex);
-            $beforeEndToken->clear();
+        // if there is some item between braces then add `,` after it
+        if ($startIndex !== $beforeEndIndex && !$beforeEndToken->equalsAny(array(',', array(T_END_HEREDOC)))) {
+            $tokens->insertAt($beforeEndIndex + 1, new Token(','));
+
+            $endToken = $tokens[$endIndex];
+
+            if (!$endToken->isComment() && !$endToken->isWhitespace()) {
+                $tokens->ensureWhitespaceAtIndex($endIndex, 1, ' ');
+            }
         }
     }
 }
