+++ tests/Fixer/ArrayNotation/WhitespaceAfterCommaInArrayFixerTest.php
@@ -10,7 +10,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace PhpCsFixer\Tests\Fixer\Symfony;
+namespace PhpCsFixer\Tests\Fixer\ArrayNotation;
 
 use PhpCsFixer\Test\AbstractFixerTestCase;
